Removing duplicate code and adding cluster_utils.go and utils.go to the /pkg dir
