+++ internal/resources/cluster.go
@@ -16,7 +16,7 @@ import (
 	"github.com/HewlettPackard/hpegl-containers-go-sdk/pkg/mcaasapi"
 
 	"github.com/HewlettPackard/hpegl-containers-terraform-resources/internal/resources/schemas"
-	"github.com/HewlettPackard/hpegl-containers-terraform-resources/internal/utils"
+	"github.com/HewlettPackard/hpegl-containers-terraform-resources/pkg/utils"
 	"github.com/HewlettPackard/hpegl-containers-terraform-resources/pkg/auth"
 	"github.com/HewlettPackard/hpegl-containers-terraform-resources/pkg/client"
 )
