+++ pkg/utils/utils.go
@@ -2,11 +2,39 @@ package utils
 
 import "github.com/HewlettPackard/hpegl-containers-go-sdk/pkg/mcaasapi"
 
-func WorkerPresentInMachineSets(machineSets []mcaasapi.MachineSet, workername string) bool {
-	for _, ms := range machineSets {
-		if ms.Name == workername {
-			return true
-		}
+func GetErrorMessage(err error, statusCode int) string {
+	swaggerErr, ok := err.(mcaasapi.GenericSwaggerError)
+	if !ok {
+		return ""
 	}
-	return false
+
+	model := swaggerErr.Model()
+	switch statusCode {
+	case 400:
+		badRequestModel, ok := model.(mcaasapi.BadRequestError)
+		if !ok {
+			return ""
+		}
+		return badRequestModel.Message
+	case 401:
+		authErrorModel, ok := model.(mcaasapi.AuthenticationError)
+		if !ok {
+			return ""
+		}
+		return authErrorModel.Message
+	case 422:
+		unprocessingEntityModel, ok := model.(mcaasapi.UnprocessingEntityError)
+		if !ok {
+			return ""
+		}
+		return unprocessingEntityModel.Message
+	case 500:
+		internalErrorModel, ok := model.(mcaasapi.InternalError)
+		if !ok {
+			return ""
+		}
+		return internalErrorModel.Message
+	default:
+		return ""
+	}
 }
